backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.9
pydantic-settings==2.5.2
Pillow==10.4.0
httpx>=0.27.0
